Refactor auth servlet to use BaseServlet and move verycommon to dev.galasa.api.common
