+++ galasa-parent/dev.galasa.framework.api.common/src/main/java/dev/galasa/framework/api/common/BaseRoute.java
@@ -3,12 +3,10 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package dev.galasa.framework.api.ras.internal.verycommon;
+package dev.galasa.framework.api.common;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-
-import dev.galasa.framework.api.common.ResponseBuilder;
 
 public abstract class BaseRoute implements IRoute {
