+++ galasa-parent/dev.galasa.framework.api.common/src/main/java/dev/galasa/framework/api/common/IRoute.java
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package dev.galasa.framework.api.ras.internal.verycommon;
+package dev.galasa.framework.api.common;
 
 import java.io.IOException;
 
@@ -15,12 +15,12 @@ import dev.galasa.framework.spi.FrameworkException;
 /**
  * IRoute provides methods for endpoints to implement when a request is sent through a servlet,
  * allowing for new routes to be added without needing servlets to know which route handles the request.
- * 
+ *
  * Route paths represent the regex patterns that are used to match request paths against.
  */
 public interface IRoute {
     String getPath();
 
-    HttpServletResponse handleRequest(String pathInfo, QueryParameters queryParams, HttpServletResponse response) 
+    HttpServletResponse handleRequest(String pathInfo, QueryParameters queryParams, HttpServletResponse response)
         throws ServletException, IOException, FrameworkException;
 }
