+++ galasa-parent/dev.galasa.framework.api.ras/src/main/java/dev/galasa/framework/api/ras/internal/RequestorRas.java
@@ -27,7 +27,7 @@ import org.osgi.service.component.annotations.Reference;
 import org.osgi.service.component.annotations.ServiceScope;
 
 import dev.galasa.framework.api.ras.internal.common.RasQueryParameters;
-import dev.galasa.framework.api.ras.internal.verycommon.QueryParameters;
+import dev.galasa.framework.api.common.QueryParameters;
 import dev.galasa.framework.spi.IFramework;
 import dev.galasa.framework.spi.IResultArchiveStoreDirectoryService;
 import dev.galasa.framework.spi.ResultArchiveStoreException;
@@ -58,7 +58,7 @@ public class RequestorRas extends HttpServlet {
 			//sorts list
 			Collections.sort(list);
 
-			JsonObject requestors = new JsonObject();   	
+			JsonObject requestors = new JsonObject();
 
 			if(!queryParameters.isAscending("requestor")) {
 				Collections.reverse(list);
