+++ galasa-parent/dev.galasa.framework.api.ras/src/test/java/dev/galasa/framework/api/ras/internal/TestQueryParameters.java
@@ -12,7 +12,7 @@ import java.util.*;
 import org.junit.Test;
 
 import dev.galasa.framework.api.common.InternalServletException;
-import dev.galasa.framework.api.ras.internal.verycommon.QueryParameters;
+import dev.galasa.framework.api.common.QueryParameters;
 
 import static org.assertj.core.api.Assertions.*;
